Feat: 스터디 채팅방 기능 추가

+++ src/main/java/com/zerobase/morse/model/StudyUpdate.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Builder
 public class StudyUpdate {
 
-	Long studyNo;
+	int studyNo;
 	String title;
 	String category;	//대분류
 	String subCategory;	//중분류
